+++ src/layouting/boundingbox.jl
@@ -1,162 +1,127 @@
-function parent_transform(x)
-    p = parent(transformation(x))
-    # TODO: Float64 model
-    isnothing(p) ? Mat4d(I) : Mat4d(p.model[])
-end
 
-function boundingbox(x, exclude = (p)-> false)
-    return parent_transform(x) * data_limits(x, exclude)
-end
+################################################################################
+### boundingbox
+################################################################################
 
-function project_widths(matrix, vec)
-    pr = project(matrix, vec)
-    zero = project(matrix, zeros(typeof(vec)))
-    return pr - zero
-end
 
-function rotate_bbox(bb::Rect3{T}, rot) where {T <: Real}
-    points = decompose(Point3{T}, bb)
-    Rect3{T}(Ref(rot) .* points)
-end
+"""
+    boundingbox(scenelike[, exclude = plot -> false])
 
-function gl_bboxes(gl::GlyphCollection)
-    scales = gl.scales.sv isa Vec2 ? (gl.scales.sv for _ in gl.extents) : gl.scales.sv
-    map(gl.glyphs, gl.extents, scales) do c, ext, scale
-        hi_bb = height_insensitive_boundingbox_with_advance(ext)
-        # TODO c != 0 filters out all non renderables, which is not always desired
-        return Rect2d(origin(hi_bb) * scale, (c != 0) * widths(hi_bb) * scale)
-    end
-end
+Returns the combined world space bounding box of all plots collected under
+`scenelike`. This include `plot.transformation`, i.e. the `transform_func` and
+the `model` matrix. Plots with `exclude(plot) == true` are excluded.
 
-function height_insensitive_boundingbox(ext::GlyphExtent)
-    l = ext.ink_bounding_box.origin[1]
-    w = ext.ink_bounding_box.widths[1]
-    b = ext.descender
-    h = ext.ascender
-    return Rect2d((l, b), (w, h - b))
-end
-
-function height_insensitive_boundingbox_with_advance(ext::GlyphExtent)
-    l = 0.0
-    r = ext.hadvance
-    b = ext.descender
-    h = ext.ascender
-    return Rect2d((l, b), (r - l, h - b))
-end
-
-_inkboundingbox(ext::GlyphExtent) = ext.ink_bounding_box
-
-function unchecked_boundingbox(glyphcollection::GlyphCollection, position::Point3, rotation::Quaternion)
-    return unchecked_boundingbox(glyphcollection, rotation) + position
-end
-
-function unchecked_boundingbox(glyphcollection::GlyphCollection, rotation::Quaternion)
-    isempty(glyphcollection.glyphs) && return Rect3d(Point3d(0), Vec3d(0))
-
-    glyphorigins = glyphcollection.origins
-    glyphbbs = gl_bboxes(glyphcollection)
-
-    bb = Rect3d()
-    for (charo, glyphbb) in zip(glyphorigins, glyphbbs)
-        charbb = rotate_bbox(Rect3d(glyphbb), rotation) + charo
-        if !isfinite_rect(bb)
-            bb = charbb
-        else
-            bb = union(bb, charbb)
+See also: [`data_limits`](@ref)
+"""
+function boundingbox(scenelike, exclude = (p)-> false)
+    bb_ref = Base.RefValue(Rect3d())
+    foreach_plot(scenelike) do plot
+        if !exclude(plot)
+            update_boundingbox!(bb_ref, boundingbox(plot))
         end
     end
-    return bb
-end
-
-function unchecked_boundingbox(layouts::AbstractArray{<:GlyphCollection}, positions, rotations)
-    isempty(layouts) && return Rect3d((0, 0, 0), (0, 0, 0))
-
-    bb = Rect3d()
-    broadcast_foreach(layouts, positions, rotations) do layout, pos, rot
-        if !isfinite_rect(bb)
-            bb = boundingbox(layout, pos, rot)
-        else
-            bb = union(bb, boundingbox(layout, pos, rot))
-        end
-    end
-    return bb
-end
-
-function boundingbox(x::Union{GlyphCollection,AbstractArray{<:GlyphCollection}}, args...)
-    bb = unchecked_boundingbox(x, args...)
-    isfinite_rect(bb) || error("Invalid text boundingbox")
-    bb
-end
-
-function boundingbox(x::Text{<:Tuple{<:GlyphCollection}})
-    if x.space[] == x.markerspace[]
-        pos = to_ndim(Point3d, x.position[], 0)
-    else
-        cam = parent_scene(x).camera
-        transformed = apply_transform(x.transformation.transform_func[], x.position[])
-        pos = Makie.project(cam, x.space[], x.markerspace[], transformed)
-    end
-    return boundingbox(x[1][], pos, to_rotation(x.rotation[]))
-end
-
-function boundingbox(x::Text{<:Tuple{<:AbstractArray{<:GlyphCollection}}})
-    if x.space[] == x.markerspace[]
-        pos = to_ndim.(Point3d, x.position[], 0)
-    else
-        cam = (parent_scene(x).camera,)
-        transformed = apply_transform(x.transformation.transform_func[], x.position[])
-        pos = Makie.project.(cam, x.space[], x.markerspace[], transformed)
-    end
-    return boundingbox(x[1][], pos, to_rotation(x.rotation[]))
-end
-
-function boundingbox(plot::Text)
-    bb = Rect3d()
-    for p in plot.plots
-        _bb = boundingbox(p)
-        if !isfinite_rect(bb)
-            bb = _bb
-        elseif isfinite_rect(_bb)
-            bb = union(bb, _bb)
-        end
-    end
-    return bb
-end
-
-_is_latex_string(x::AbstractVector{<:LaTeXString}) = true
-_is_latex_string(x::LaTeXString) = true
-_is_latex_string(other) = false
-
-function text_bb(str, font, size)
-    rot = Quaternionf(0,0,0,1)
-    fonts = nothing # TODO: remove the arg if possible
-    layout = layout_text(
-        str, size, font, fonts, Vec2f(0), rot, 0.5, 1.0,
-        RGBAf(0, 0, 0, 0), RGBAf(0, 0, 0, 0), 0f0, 0f0)
-    return boundingbox(layout, Point3d(0), rot)
+    return bb_ref[]
 end
 
 """
-Calculate an approximation of a tight rectangle around a 2D rectangle rotated by `angle` radians.
-This is not perfect but works well enough. Check an A vs X to see the difference.
+    boundingbox(plot::AbstractPlot)
+
+Returns the world space bounding box of a plot. This include `plot.transformation`,
+i.e. the `transform_func` and the `model` matrix.
+
+See also: [`data_limits`](@ref)
 """
-function rotatedrect(rect::Rect{2, T}, angle)::Rect{2, T} where T
-    ox, oy = rect.origin
-    wx, wy = rect.widths
-    points = Mat{2, 4, T}(
-        ox, oy,
-        ox, oy+wy,
-        ox+wx, oy,
-        ox+wx, oy+wy
-    )
-    mrot = Mat{2, 2, T}(
-        cos(angle), -sin(angle),
-        sin(angle), cos(angle)
-    )
-    rotated = mrot * points
+boundingbox(plot::AbstractPlot) = _boundingbox(plot)
 
-    rmins = minimum(rotated; dims=2)
-    rmaxs = maximum(rotated; dims=2)
+# TODO: This only exists to deprecate boundingbox(::Text) more smoothly. Once
+#       that is fully removed this should be boundingbox(plot).
+function _boundingbox(plot::AbstractPlot)
+    # Assume primitive plot
+    if isempty(plot.plots)
+        return Rect3d(iterate_transformed(plot))
+    end
 
-    return Rect2(rmins..., (rmaxs .- rmins)...)
+    # Assume combined plot
+    bb_ref = Base.RefValue(boundingbox(plot.plots[1]))
+    for i in 2:length(plot.plots)
+        update_boundingbox!(bb_ref, boundingbox(plot.plots[i]))
+    end
+
+    return
+end
+
+# for convenience
+function _boundingbox(plot::AbstractPlot, lims::Rect)
+    return Rect3d(iterate_transformed(plot, point_iterator(lims)))
+end
+
+
+
+################################################################################
+### transformed point iterator
+################################################################################
+
+
+@inline iterate_transformed(plot) = iterate_transformed(plot, point_iterator(plot))
+
+function iterate_transformed(plot, points)
+    t = transformation(plot)
+    model = Mat4d(model_transform(t)) # TODO: make model matrix Float64?
+    trans_func = transform_func(t)
+    return iterate_transformed(points, model, to_value(get(plot, :space, :data)), trans_func)
+end
+
+function iterate_transformed(points::AbstractVector{<: VecTypes{N, T}}, model, space, trans_func) where {N, T}
+    output = similar(points, Point3d)
+    @inbounds for i in eachindex(points)
+        transformed = apply_transform(trans_func, points[i], space)
+        p4d = project(model, transformed)
+        output[i] = p4d[Vec(1, 2, 3)]
+    end
+    return output
+end
+
+function iterate_transformed(points::T, model, space, trans_func) where T
+    @warn "iterate_transformed with $T"
+    [to_ndim(Point3f, project(model, apply_transform(trans_func, point, space))) for point in points]
+end
+
+
+################################################################################
+### Special cases
+################################################################################
+
+
+# includes markersize and rotation
+function boundingbox(plot::MeshScatter)
+    # TODO: avoid mesh generation here if possible
+    @get_attribute plot (marker, markersize, rotations)
+    marker_bb = Rect3d(marker)
+    positions = iterate_transformed(plot)
+    scales = markersize
+    # fast path for constant markersize
+    if scales isa VecTypes{3} && rotations isa Quaternion
+        bb = Rect3d(positions)
+        marker_bb = rotations * (marker_bb * scales)
+        return Rect3d(minimum(bb) + minimum(marker_bb), widths(bb) + widths(marker_bb))
+    else
+        # TODO: optimize const scale, var rot and var scale, const rot
+        return limits_from_transformed_points(positions, scales, rotations, marker_bb)
+    end
+end
+
+# include bbox from scaled markers
+function limits_from_transformed_points(positions, scales, rotations, element_bbox)
+    isempty(positions) && return Rect3d()
+
+    first_scale = attr_broadcast_getindex(scales, 1)
+    first_rot = attr_broadcast_getindex(rotations, 1)
+    full_bbox = Ref(first_rot * (element_bbox * first_scale) + first(positions))
+    for (i, pos) in enumerate(positions)
+        scale, rot = attr_broadcast_getindex(scales, i), attr_broadcast_getindex(rotations, i)
+        transformed_bbox = rot * (element_bbox * scale) + pos
+        update_boundingbox!(full_bbox, transformed_bbox)
+    end
+
+    return full_bbox[]
 end
