+++ src/layouting/data_limits.jl
@@ -1,10 +1,111 @@
-_isfinite(x) = isfinite(x)
-_isfinite(x::VecTypes) = all(isfinite, x)
+#=
+Hierarchy:
+- boundingbox falls back on points_iterator of primitives
+- points_iterator falls back on data_limits
+- data_limits uses points_iterator for a few specific primitive plots
+
+So overload both `data_limits` and `boundingbox`. You can use:
+- `points_iterator(::Rect)` to decompose the Rect
+- `_boundingbox(plot, ::Rect)` to transform the Rect using the plots transformations
+=#
+
+################################################################################
+### data_limits
+################################################################################
+
+"""
+    data_limits(scenelike[, exclude = plot -> false])
+
+Returns the combined data limits of all plots collected under `scenelike` for
+which `exclude(plot) == false`. This is solely based on the positional data of
+a plot and thus does not include any transformations.
+
+See also: [`boundingbox`](@ref)
+"""
+function data_limits(scenelike, exclude=(p)-> false)
+    bb_ref = Base.RefValue(Rect3d())
+    foreach_plot(scenelike) do plot
+        if !exclude(plot)
+            update_boundingbox!(bb_ref, data_limits(plot))
+        end
+    end
+    return bb_ref[]
+end
+
+"""
+    data_limits(plot::AbstractPlot)
+
+Returns the bounding box of a plot based on just its position data.
+
+See also: [`boundingbox`](@ref)
+"""
+function data_limits(plot::AbstractPlot)
+    # Assume primitive plot
+    if isempty(plot.plots)
+        return Rect3d(point_iterator(plot))
+    end
+
+    # Assume combined plot
+    bb_ref = Base.RefValue(data_limits(plot.plots[1]))
+    for i in 2:length(plot.plots)
+        update_boundingbox!(bb_ref, data_limits(plot.plots[i]))
+    end
+
+    return bb_ref[]
+end
+
+# A few overloads for performance
+function data_limits(plot::Surface)
+    mini_maxi = extrema_nan.((plot.x[], plot.y[], plot.z[]))
+    mini = first.(mini_maxi)
+    maxi = last.(mini_maxi)
+    return Rect3d(mini, maxi .- mini)
+end
+
+function data_limits(plot::Union{Heatmap, Image})
+    mini_maxi = extrema_nan.((plot.x[], plot.y[]))
+    mini = Vec3d(first.(mini_maxi)..., 0)
+    maxi = Vec3d(last.(mini_maxi)..., 0)
+    return Rect3d(mini, maxi .- mini)
+end
+
+function data_limits(x::Volume)
+    axes = (x[1][], x[2][], x[3][])
+    extremata = extrema.(axes)
+    return Rect3d(first.(extremata), last.(extremata) .- first.(extremata))
+end
+
+
+################################################################################
+### point_iterator & data_limits
+################################################################################
+
+
+function point_iterator(plot::Union{Scatter, MeshScatter, Lines, LineSegments})
+    return plot.positions[]
+end
+
+point_iterator(plot::Text) = point_iterator(plot.plots[1])
+function point_iterator(plot::Text{<: Tuple{<: Union{GlyphCollection, AbstractVector{GlyphCollection}}}})
+    return plot.positions[]
+end
+
+point_iterator(mesh::GeometryBasics.Mesh) = decompose(Point, mesh)
+point_iterator(plot::Mesh) = point_iterator(plot.mesh[])
+
+# Fallback for other primitive plots, used in boundingbox
+point_iterator(plot::AbstractPlot) = points_iterator(data_limits(plot))
+
+# For generic usage
+point_iterator(bbox::Rect) = unique(decompose(Point3d, bbox))
+
+
+################################################################################
+### Utilities
+################################################################################
+
+
 isfinite_rect(x::Rect) = all(isfinite, x.origin) &&  all(isfinite, x.widths)
-scalarmax(x::Union{Tuple, AbstractArray}, y::Union{Tuple, AbstractArray}) = max.(x, y)
-scalarmax(x, y) = max(x, y)
-scalarmin(x::Union{Tuple, AbstractArray}, y::Union{Tuple, AbstractArray}) = min.(x, y)
-scalarmin(x, y) = min(x, y)
 
 extrema_nan(itr::Pair) = (itr[1], itr[2])
 extrema_nan(itr::ClosedInterval) = (minimum(itr), maximum(itr))
@@ -30,133 +131,6 @@ function extrema_nan(itr)
     return (vmin, vmax)
 end
 
-function distinct_extrema_nan(x)
-    lo, hi = extrema_nan(x)
-    lo == hi ? (lo - 0.5f0, hi + 0.5f0) : (lo, hi)
-end
-
-function point_iterator(plot::Union{Scatter, MeshScatter, Lines, LineSegments})
-    return plot.positions[]
-end
-
-# TODO?
-function data_limits(text::Text{<: Tuple{<: Union{GlyphCollection, AbstractVector{GlyphCollection}}}})
-    if is_data_space(text.markerspace[])
-        return boundingbox(text)
-    else
-        if text.position[] isa VecTypes
-            return Rect3d(text.position[])
-        else
-            # TODO: is this branch necessary?
-            return Rect3d(convert_arguments(PointBased(), text.position[])[1])
-        end
-    end
-end
-
-function data_limits(text::Text)
-    return data_limits(text.plots[1])
-end
-
-point_iterator(mesh::GeometryBasics.Mesh) = decompose(Point, mesh)
-
-function point_iterator(list::AbstractVector)
-    if length(list) == 1
-        # save a copy!
-        return point_iterator(list[1])
-    else
-        points = Point3d[]
-        for elem in list
-            for point in point_iterator(elem)
-                push!(points, to_ndim(Point3d, point, 0))
-            end
-        end
-        return points
-    end
-end
-
-point_iterator(plot::Mesh) = point_iterator(plot.mesh[])
-
-function br_getindex(vector::AbstractVector, idx::CartesianIndex, dim::Int)
-    return vector[Tuple(idx)[dim]]
-end
-
-function br_getindex(matrix::AbstractMatrix, idx::CartesianIndex, dim::Int)
-    return matrix[idx]
-end
-
-function point_iterator(plot::Union{Image, Heatmap, Surface})
-    rect = data_limits(plot)
-    return unique(decompose(Point3d, rect))
-end
-
-function point_iterator(x::Volume)
-    axes = (x[1], x[2], x[3])
-    extremata = map(extrema∘to_value, axes)
-    minpoint = Point3d(first.(extremata)...)
-    widths = last.(extremata) .- first.(extremata)
-    rect = Rect3d(minpoint, Vec3d(widths))
-    return unique(decompose(Point, rect))
-end
-
-function foreach_plot(f, s::Scene)
-    foreach_plot(f, s.plots)
-    foreach(sub-> foreach_plot(f, sub), s.children)
-end
-
-foreach_plot(f, s::Figure) = foreach_plot(f, s.scene)
-foreach_plot(f, s::FigureAxisPlot) = foreach_plot(f, s.figure)
-foreach_plot(f, list::AbstractVector) = foreach(f, list)
-function foreach_plot(f, plot::Plot)
-    if isempty(plot.plots)
-        f(plot)
-    else
-        foreach_plot(f, plot.plots)
-    end
-end
-
-function foreach_transformed(f, point_iterator, model, trans_func)
-    for point in point_iterator
-        point_t = apply_transform(trans_func, point)
-        point_m = project(model, point_t)
-        f(point_m)
-    end
-    return
-end
-
-function foreach_transformed(f, plot)
-    points = point_iterator(plot)
-    t = transformation(plot)
-    model = model_transform(t)
-    trans_func = t.transform_func[]
-    # use function barrier since trans_func is Any
-    foreach_transformed(f, points, model, identity)
-end
-
-# TODO:
-# These two functions currently:
-# - skip transfunc
-# - skip translations from model (due to 0f0 in project)
-# - apply scaling and rotation
-# What do we actually want from them?
-function iterate_transformed(plot)
-    points = point_iterator(plot)
-    t = transformation(plot)
-    model = model_transform(t)
-    # TODO: without this, axes with log scales error.  Why?
-    trans_func = identity # transform_func(t)
-    # trans_func = identity
-    iterate_transformed(points, model, to_value(get(plot, :space, :data)), trans_func)
-end
-function iterate_transformed(points::AbstractVector{<: VecTypes{N, T}}, model, space, trans_func) where {N, T}
-    # TODO: either skip this all-together or make model a Float64 matrix
-    [to_ndim(Point3d, project(Mat4{T}(model), apply_transform(trans_func, point, space)), zero(T)) for point in points]
-end
-# TODO: this is just to catch types the above misses
-function iterate_transformed(points::T, model, space, trans_func) where T
-    @warn "iterate_transformed with $T"
-    [to_ndim(Point3f, project(model, apply_transform(trans_func, point, space)), 0f0) for point in points]
-end
-
 function update_boundingbox!(bb_ref, point)
     if all(isfinite, point)
         vec = to_ndim(Vec3d, point, 0.0)
@@ -177,103 +151,19 @@ function update_boundingbox!(bb_ref, bb::Rect)
     return
 end
 
-# Default data_limits
-function data_limits(plot::AbstractPlot)
-    # Assume primitive plot
+
+function foreach_plot(f, s::Scene)
+    foreach_plot(f, s.plots)
+    foreach(sub-> foreach_plot(f, sub), s.children)
+end
+
+foreach_plot(f, s::Figure) = foreach_plot(f, s.scene)
+foreach_plot(f, s::FigureAxisPlot) = foreach_plot(f, s.figure)
+foreach_plot(f, list::AbstractVector) = foreach(f, list)
+function foreach_plot(f, plot::Plot)
     if isempty(plot.plots)
-        return limits_from_transformed_points(iterate_transformed(plot))
-    end
-
-    # Assume Plot Plot
-    bb_ref = Base.RefValue(data_limits(plot.plots[1]))
-    for i in 2:length(plot.plots)
-        update_boundingbox!(bb_ref, data_limits(plot.plots[i]))
-    end
-
-    return bb_ref[]
-end
-
-function _update_rect(rect::Rect{N, T}, point::VecTypes{N, T}) where {N, T}
-    mi = minimum(rect)
-    ma = maximum(rect)
-    mis_mas = map(mi, ma, point) do _mi, _ma, _p
-        (isnan(_mi) ? _p : _p < _mi ? _p : _mi), (isnan(_ma) ? _p : _p > _ma ? _p : _ma)
-    end
-    new_o = map(first, mis_mas)
-    new_w = map(mis_mas) do (mi, ma)
-        ma - mi
-    end
-    typeof(rect)(new_o, new_w)
-end
-
-function limits_from_transformed_points(points_iterator)
-    isempty(points_iterator) && return Rect3d()
-    first, rest = Iterators.peel(points_iterator)
-    bb = foldl(_update_rect, rest, init = Rect3d(first, zero(first)))
-    return bb
-end
-
-# include bbox from scaled markers
-function limits_from_transformed_points(positions, scales, rotations, element_bbox)
-    isempty(positions) && return Rect3d()
-
-    first_scale = attr_broadcast_getindex(scales, 1)
-    first_rot = attr_broadcast_getindex(rotations, 1)
-    full_bbox = Ref(first_rot * (element_bbox * first_scale) + first(positions))
-    for (i, pos) in enumerate(positions)
-        scale, rot = attr_broadcast_getindex(scales, i), attr_broadcast_getindex(rotations, i)
-        transformed_bbox = rot * (element_bbox * scale) + pos
-        update_boundingbox!(full_bbox, transformed_bbox)
-    end
-
-    return full_bbox[]
-end
-
-function data_limits(scenelike, exclude=(p)-> false)
-    bb_ref = Base.RefValue(Rect3d())
-    foreach_plot(scenelike) do plot
-        if !exclude(plot)
-            update_boundingbox!(bb_ref, data_limits(plot))
-        end
-    end
-    return bb_ref[]
-end
-
-# A few overloads for performance
-function data_limits(plot::Surface)
-    mini_maxi = extrema_nan.((plot.x[], plot.y[], plot.z[]))
-    mini = first.(mini_maxi)
-    maxi = last.(mini_maxi)
-    return Rect3d(mini, maxi .- mini)
-end
-
-function data_limits(plot::Heatmap)
-    mini_maxi = extrema_nan.((plot.x[], plot.y[]))
-    mini = Vec3d(first.(mini_maxi)..., 0)
-    maxi = Vec3d(last.(mini_maxi)..., 0)
-    return Rect3d(mini, maxi .- mini)
-end
-
-function data_limits(plot::Image)
-    mini_maxi = extrema_nan.((plot.x[], plot.y[]))
-    mini = Vec3d(first.(mini_maxi)..., 0)
-    maxi = Vec3d(last.(mini_maxi)..., 0)
-    return Rect3d(mini, maxi .- mini)
-end
-
-function data_limits(plot::MeshScatter)
-    # TODO: avoid mesh generation here if possible
-    @get_attribute plot (marker, markersize, rotations)
-    marker_bb = Rect3d(marker)
-    positions = iterate_transformed(plot)
-    scales = markersize
-    # fast path for constant markersize
-    if scales isa VecTypes{3} && rotations isa Quaternion
-        bb = limits_from_transformed_points(positions)
-        marker_bb = rotations * (marker_bb * scales)
-        return Rect3d(minimum(bb) + minimum(marker_bb), widths(bb) + widths(marker_bb))
+        f(plot)
     else
-        # TODO: optimize const scale, var rot and var scale, const rot
-        return limits_from_transformed_points(positions, scales, rotations, marker_bb)
+        foreach_plot(f, plot.plots)
     end
 end
